requirements.txt
# Web UI
streamlit==1.32.2

# Language Models and Chains
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.32
langchain-google-genai==0.0.5  # Gemini API integration

# Embeddings
huggingface-hub==0.20.3
sentence-transformers==2.6.1

# Vector DB
chromadb==0.4.24

# Document Loaders & Parsers
pypdf==3.17.4
python-docx==1.1.0
python-pptx==0.6.23
pandas==2.2.1
openpyxl==3.1.2
docx2txt==0.8

# Env variables
python-dotenv==1.0.1

# Misc
tqdm==4.66.2
